% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eudata.R
\name{eudata_dataset}
\alias{eudata_dataset}
\title{Get details for a specific EU dataset}
\usage{
eudata_dataset(id)
}
\arguments{
\item{id}{Dataset identifier (e.g. "groenklimaatassen-gent")}
}
\value{
tibble with one row: id, title, description, country, catalog,
publisher, modified, issued, format, resource_url
}
\description{
Get details for a specific EU dataset
}
